Added EndPoints catalogItemsV0.md
